Move redundant CLI arguments to a shared module
Adds new exlucde argument to delocate-path

Makes verbosity work the same among all commands
